Fix a collection of compiler warnings across OPAL and OMPI
Unused variables:
  ompi/communicator/comm_cid.c: remove unused 'comm' local
  ompi/mca/coll/han/coll_han_subcomms.c: remove unused 'w_rank' in two
    functions

Unsigned-enum comparisons that are always true/false:
  Enum types used as flags are unsigned in practice; comparisons of the
  form 'value < 0' can never be true and are removed:
    ompi/mca/coll/han/coll_han_dynamic.c (COMPONENT_T)
    ompi/mca/coll/tuned/coll_tuned_dynamic_rules.c (comm_rank_distro_t)
    ompi/mca/coll/xhc/coll_xhc_component.c (XHC_COLLTYPE_T,
      xhc_config_source_t)
    opal/mca/base/mca_base_event.c (MCA_BASE_VAR_TYPE)
    opal/mca/base/mca_base_event_dispatch.c (MCA_BASE_EVENT_CB_REQUIRE_T)

Format-specifier mismatches:
  ompi/communicator/comm.c: %ld -> PRIu64 for uint64_t
  ompi/mca/coll/han/coll_han_alltoallv.c: %ld -> PRId64 for int64_t
    (four call sites)
  ompi/mca/coll/tuned/coll_tuned_dynamic_file.c: %ld -> PRId64 for
    int64_t

VLA-folded-to-constant:
  ompi/mca/coll/han/coll_han_alltoallv.c: replace 'const int N = 8'
  array bound with 'enum { N = 8 }' to give a true integer constant
  expression and avoid the -Wvla warning (two functions).

Sign-compare:
  ompi/mpi/fortran/base/strings.c: cast strlen() results to int when
  comparing against the Fortran hidden-length argument (which is int).

+++ ompi/mpi/fortran/base/strings.c
@@ -101,12 +101,12 @@ int ompi_fortran_string_c2f(const char *cstr, char *fstr, int len)
     // trailing \0 into the last position in fstr.  This is not what
     // Fortran wants; overwrite that \0 with the actual last character
     // that will fit into fstr.
-    if (len <= strlen(cstr)) {
+    if (len <= (int) strlen(cstr)) {
         fstr[len - 1] = cstr[len - 1];
     } else {
         // Otherwise, pad the end of the resulting Fortran string with
         // spaces.
-        for (i = strlen(cstr); i < len; ++i) {
+        for (i = (int) strlen(cstr); i < len; ++i) {
             fstr[i] = ' ';
         }
     }
